Applying refactoring changes from PR review
