+++ src/imageLoader/wadors/metaData/metaDataProvider.js
@@ -9,12 +9,12 @@ import {
   getMultiframeInformation,
   getFrameInformation,
 } from '../combineFrameInstance.js';
+import { retrieveFirstFrameMetadata } from '../retrieveFirstFrameMetadata.js';
 
 function metaDataProvider(type, imageId) {
   if (type === 'MultiframeModule') {
     // the get function removes the PerFrameFunctionalGroupsSequence
-    const { metadata, frame } =
-      metaDataManager.retrieveFirstFrameMetadata(imageId);
+    const { metadata, frame } = retrieveFirstFrameMetadata(imageId);
 
     if (!metadata) {
       return;
