+++ src/imageLoader/wadouri/dataSetCacheManager.js
@@ -15,75 +15,9 @@ let loadedDataSets = {};
 
 let promises = {};
 
-function isMultiFrame(uri) {
-  const dataSet = _get(uri);
-
-  return _isMultiFrame(dataSet);
-}
-
-function _isMultiFrame(dataSet) {
-  // Checks if dicomTag NumberOf Frames exists and it is greater than one
-  if (!dataSet) {
-    return false;
-  }
-
-  const numberOfFrames = dataSet.intString('x00280008');
-
-  return numberOfFrames && numberOfFrames > 1;
-}
-
-function retrieveFrameParameterIndex(uri) {
-  return uri.indexOf('&frame=');
-}
-
-function _get(uri) {
-  if (!loadedDataSets[uri]) {
-    return;
-  }
-
-  return loadedDataSets[uri];
-}
-
-function retrieveFirstFrameMetadata(uri) {
-  const frameParameterIndex = retrieveFrameParameterIndex(uri);
-  const multiframeURI = uri.slice(0, frameParameterIndex);
-  const frame = parseInt(uri.slice(frameParameterIndex + 7), 10);
-
-  let dataSet;
-
-  if (loadedDataSets[uri]) {
-    dataSet = loadedDataSets[multiframeURI].dataSet;
-  } else {
-    dataSet = undefined;
-  }
-
-  return {
-    dataSet,
-    frame,
-  };
-}
-
 // returns true if the wadouri for the specified index has been loaded
 function isLoaded(uri) {
   return loadedDataSets[uri] !== undefined;
-}
-
-function generateMultiframeWADOURIs(uri) {
-  const listWADOURIs = [];
-
-  const dataSet = _get(uri);
-
-  if (_isMultiFrame(dataSet)) {
-    const numberOfFrames = dataSet.intString('x00280008');
-
-    for (let i = 1; i <= numberOfFrames; i++) {
-      listWADOURIs.push(`${uri}&frame=${i}`);
-    }
-  } else {
-    listWADOURIs.push(uri);
-  }
-
-  return listWADOURIs;
 }
 
 function get(uri) {
@@ -262,6 +196,8 @@ function purge() {
   cacheSizeInBytes = 0;
 }
 
+export { loadedDataSets };
+
 export default {
   isLoaded,
   load,
@@ -270,7 +206,4 @@ export default {
   purge,
   get,
   update,
-  generateMultiframeWADOURIs,
-  retrieveFirstFrameMetadata,
-  isMultiFrame,
 };
